--- src/Component/Toggle/index.js
@@ -1,1 +1,0 @@
-export {UiChangeBtn} from './UiChange';
