--- src/Component/Vending_machine/index.js
@@ -1,1 +1,0 @@
-export {VendingMachine} from './VendingMachine';
