--- src/Component/Wallet/index.js
@@ -1,1 +1,0 @@
-export {Wallet} from './Wallet';
