+++ src/Component/index.js
@@ -1,0 +1,4 @@
+export {InsertButton} from './InsertButton';
+export {Products} from './Products';
+export {UiChangeBtn} from './UiChange';
+export {VendingMachineInterface} from './VendingMachineInterface';
