+++ src/Pages/Home.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {Outlet} from 'react-router-dom';
 
-import {UiChangeBtn} from '../Component/Toggle';
+import {UiChangeBtn} from '../Component';
 
 export const Home = () => {
   return (
